Fix dll name in import library for mingw
In mingw, the DLL name is libzimg-2.dll. So, the import library
libzimg-2.dll.a should point to that DLL. With the LIBRARY name
in zimg.def, the import library points to z.dll. This removes
LIBRARY name in zimg.def file and the correct DLL name is
automatically added.

In case of MSVC, the /OUT linker option is used to specify the DLL's
output name[1].

[1]: https://learn.microsoft.com/en-us/cpp/build/reference/library